+++ node/jaeger/Cargo.toml
@@ -11,6 +11,7 @@ mick-jaeger = "0.1.4"
 lazy_static = "1.4"
 parking_lot = "0.11.1"
 polkadot-primitives = { path = "../../primitives" }
+polkadot-node-primitives = { path = "../primitives" }
 sc-network = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
 thiserror = "1.0.23"
