+++ node/jaeger/src/spans.rs
@@ -84,7 +84,8 @@
 //! ```
 
 use parity_scale_codec::Encode;
-use polkadot_primitives::v1::{BlakeTwo256, CandidateHash, Hash, HashT, Id as ParaId, PoV, ValidatorIndex};
+use polkadot_primitives::v1::{BlakeTwo256, CandidateHash, Hash, HashT, Id as ParaId, ValidatorIndex};
+use polkadot_node_primitives::PoV;
 use sc_network::PeerId;
 
 use std::fmt;
