+++ node/network/collator-protocol/src/collator_side.rs
@@ -22,8 +22,8 @@ use futures::{select, FutureExt, channel::oneshot};
 use sp_core::Pair;
 
 use polkadot_primitives::v1::{
-	CandidateHash, CandidateReceipt, CollatorPair, CompressedPoV, CoreIndex, CoreState, Hash,
-	Id as ParaId, PoV, ValidatorId
+	CandidateHash, CandidateReceipt, CollatorPair, CoreIndex, CoreState, Hash,
+	Id as ParaId, ValidatorId
 };
 use polkadot_subsystem::{
 	jaeger, PerLeafSpan,
@@ -42,7 +42,7 @@ use polkadot_node_subsystem_util::{
 	request_availability_cores_ctx,
 	metrics::{self, prometheus},
 };
-use polkadot_node_primitives::{SignedFullStatement, Statement};
+use polkadot_node_primitives::{SignedFullStatement, Statement, PoV, CompressedPoV};
 
 #[derive(Clone, Default)]
 pub struct Metrics(Option<MetricsInner>);
@@ -949,9 +949,10 @@ mod tests {
 	};
 	use polkadot_node_subsystem_util::TimeoutExt;
 	use polkadot_primitives::v1::{
-		AuthorityDiscoveryId, BlockData, CandidateDescriptor, CollatorPair, GroupRotationInfo,
+		AuthorityDiscoveryId, CandidateDescriptor, CollatorPair, GroupRotationInfo,
 		ScheduledCore, SessionIndex, SessionInfo, ValidatorIndex,
 	};
+	use polkadot_node_primitives::BlockData;
 	use polkadot_subsystem::{
 		jaeger,
 		messages::{RuntimeApiMessage, RuntimeApiRequest},
