+++ node/network/protocol/src/request_response/v1.rs
@@ -19,10 +19,11 @@
 use parity_scale_codec::{Decode, Encode};
 
 use polkadot_primitives::v1::{
-	AvailableData, CandidateHash, CandidateReceipt, ErasureChunk, ValidatorIndex,
-	CompressedPoV, Hash,
+	CandidateHash, CandidateReceipt, ValidatorIndex,
+	Hash,
 };
 use polkadot_primitives::v1::Id as ParaId;
+use polkadot_node_primitives::{AvailableData, CompressedPoV, ErasureChunk};
 
 use super::request::IsRequest;
 use super::Protocol;
